requirement_approval/requirements.txt
Django==4.2.0
mysqlclient==2.2.0
python-decouple==3.8
Pillow==10.0.0
reportlab==4.0.4
celery==5.3.0
redis==5.0.0
weasyprint==59.0
